refactor(tui): simplify bubble tea implementation
