+++ src/presentation/ui/tui/internal/bubble_tea/configurator_server.go
@@ -1,0 +1,201 @@
+package bubble_tea
+
+import (
+	"fmt"
+	"strings"
+
+	appConfiguration "tungo/application/configuration"
+	"tungo/application/runtime"
+
+	tea "charm.land/bubbletea/v2"
+)
+
+func (m configuratorSessionModel) updateServerSelectScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
+	switch msg.String() {
+	case "esc":
+		m.notice = ""
+		m.cursor = 0
+		m.screen = configuratorScreenMode
+		return m, nil
+	}
+
+	m.updateCursor(msg, len(m.server.menuOptions))
+	if msg.String() != "enter" || len(m.server.menuOptions) == 0 {
+		return m, nil
+	}
+
+	switch m.server.menuOptions[m.cursor] {
+	case sessionServerStart:
+		m = m.startModeWithDaemonGuard(runtime.ModeServer, configuratorScreenServerSelect, false)
+		if m.done {
+			return m, tea.Quit
+		}
+		return m, nil
+	case sessionServerAdd:
+		generated, err := m.options.ServerConfigurationControl.GenerateClientConfiguration()
+		if err != nil {
+			m.resultErr = err
+			m.done = true
+			return m, tea.Quit
+		}
+		m.notice = fmt.Sprintf("Client configuration saved to %s", generated.Path)
+		return m, nil
+	case sessionServerManage:
+		peers, err := m.options.ServerConfigurationControl.ListPeers()
+		if err != nil {
+			m.resultErr = err
+			m.done = true
+			return m, tea.Quit
+		}
+		if len(peers) == 0 {
+			m.notice = "No clients configured yet."
+			return m, nil
+		}
+		m.server.managePeers = peers
+		m.server.manageLabels = buildServerManageLabels(peers)
+		m.notice = ""
+		m.cursor = 0
+		m.screen = configuratorScreenServerManage
+		return m, nil
+	}
+	return m, nil
+}
+
+func (m configuratorSessionModel) updateServerManageScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
+	switch msg.String() {
+	case "esc":
+		m.notice = ""
+		m.cursor = 0
+		m.screen = configuratorScreenServerSelect
+		return m, nil
+	case "d", "D":
+		if len(m.server.managePeers) == 0 {
+			return m, nil
+		}
+		m.server.deletePeer = m.server.managePeers[m.cursor]
+		m.server.deleteCursor = m.cursor
+		m.cursor = 0
+		m.screen = configuratorScreenServerDeleteConfirm
+		return m, nil
+	}
+
+	m.updateCursor(msg, len(m.server.managePeers))
+	if msg.String() != "enter" || len(m.server.managePeers) == 0 {
+		return m, nil
+	}
+
+	peer := m.server.managePeers[m.cursor]
+	nextEnabled := !peer.Enabled
+	if err := m.options.ServerConfigurationControl.SetPeerEnabled(peer.ClientID, nextEnabled); err != nil {
+		m.notice = fmt.Sprintf("Failed to update client #%d: %v", peer.ClientID, err)
+		m.screen = configuratorScreenServerSelect
+		m.cursor = 0
+		return m, nil
+	}
+
+	peers, err := m.options.ServerConfigurationControl.ListPeers()
+	if err != nil {
+		m.resultErr = err
+		m.done = true
+		return m, tea.Quit
+	}
+	if len(peers) == 0 {
+		m.notice = "No clients configured yet."
+		m.screen = configuratorScreenServerSelect
+		m.cursor = 0
+		return m, nil
+	}
+
+	m.server.managePeers = peers
+	m.server.manageLabels = buildServerManageLabels(peers)
+	if m.cursor >= len(m.server.managePeers) {
+		m.cursor = len(m.server.managePeers) - 1
+	}
+	return m, nil
+}
+
+func (m configuratorSessionModel) updateServerDeleteConfirmScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
+	switch msg.String() {
+	case "esc":
+		if len(m.server.managePeers) > 0 {
+			m.cursor = minInt(m.server.deleteCursor, len(m.server.managePeers)-1)
+		} else {
+			m.cursor = 0
+		}
+		m.screen = configuratorScreenServerManage
+		return m, nil
+	}
+
+	options := []string{sessionServerDeleteConfirm, sessionCancel}
+	m.updateCursor(msg, len(options))
+	if msg.String() != "enter" {
+		return m, nil
+	}
+
+	selected := options[m.cursor]
+	if selected == sessionCancel {
+		if len(m.server.managePeers) > 0 {
+			m.cursor = minInt(m.server.deleteCursor, len(m.server.managePeers)-1)
+		} else {
+			m.cursor = 0
+		}
+		m.screen = configuratorScreenServerManage
+		return m, nil
+	}
+
+	if err := m.options.ServerConfigurationControl.RemovePeer(m.server.deletePeer.ClientID); err != nil {
+		m.notice = fmt.Sprintf("Failed to remove client #%d: %v", m.server.deletePeer.ClientID, err)
+		m.screen = configuratorScreenServerManage
+		m.cursor = 0
+		return m, nil
+	}
+
+	peers, err := m.options.ServerConfigurationControl.ListPeers()
+	if err != nil {
+		m.resultErr = err
+		m.done = true
+		return m, tea.Quit
+	}
+	if len(peers) == 0 {
+		m.notice = "No clients configured yet."
+		m.screen = configuratorScreenServerSelect
+		m.cursor = 0
+		return m, nil
+	}
+
+	m.notice = fmt.Sprintf(
+		"Client #%d %s removed.",
+		m.server.deletePeer.ClientID,
+		serverPeerDisplayName(m.server.deletePeer),
+	)
+	m.server.managePeers = peers
+	m.server.manageLabels = buildServerManageLabels(peers)
+	m.cursor = minInt(m.server.deleteCursor, len(peers)-1)
+	m.screen = configuratorScreenServerManage
+	return m, nil
+}
+
+func buildServerManageLabels(peers []appConfiguration.ServerPeer) []string {
+	labels := make([]string, 0, len(peers))
+	for _, peer := range peers {
+		labels = append(labels, serverPeerOptionLabel(peer))
+	}
+	return labels
+}
+
+func serverPeerDisplayName(peer appConfiguration.ServerPeer) string {
+	name := strings.TrimSpace(peer.Name)
+	if name == "" {
+		return fmt.Sprintf("client-%d", peer.ClientID)
+	}
+	return name
+}
+
+func serverPeerOptionLabel(peer appConfiguration.ServerPeer) string {
+	status := "disabled"
+	if peer.Enabled {
+		status = "enabled"
+	}
+	name := serverPeerDisplayName(peer)
+	return fmt.Sprintf("#%d %s [%s]", peer.ClientID, name, status)
+}
