+++ src/presentation/ui/tui/internal/bubble_tea/configurator_session.go
@@ -1,13 +1,11 @@
 package bubble_tea
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"slices"
 	"strings"
 	"time"
-	"unicode/utf8"
 
 	appConfiguration "tungo/application/configuration"
 	"tungo/application/runtime"
@@ -31,14 +29,6 @@ const (
 	configuratorTabSettings
 	configuratorTabLogs
 )
-
-type configuratorSessionProgram interface {
-	Run() (tea.Model, error)
-}
-
-var newConfiguratorSessionProgram = func(model tea.Model) configuratorSessionProgram {
-	return tea.NewProgram(model)
-}
 
 type ConfiguratorSessionOptions struct {
 	ClientConfigurationControl appConfiguration.ClientConfigurationControl
@@ -160,31 +150,6 @@ type configuratorSessionModel struct {
 	done       bool
 }
 
-func RunConfiguratorSession(options ConfiguratorSessionOptions) (selectedMode runtime.Mode, err error) {
-	defer clearTerminalAfterTUI()
-
-	settings := loadUISettingsFromDisk()
-	model, err := newConfiguratorSessionModel(options, settings)
-	if err != nil {
-		return 0, err
-	}
-
-	program := newConfiguratorSessionProgram(model)
-	finalModel, runErr := program.Run()
-	if runErr != nil {
-		return 0, runErr
-	}
-
-	result, ok := finalModel.(configuratorSessionModel)
-	if !ok {
-		return 0, errors.New("invalid configurator session model")
-	}
-	if result.resultErr != nil {
-		return 0, result.resultErr
-	}
-	return result.resultMode, nil
-}
-
 func newConfiguratorSessionModel(options ConfiguratorSessionOptions, settings *uiPreferencesProvider) (configuratorSessionModel, error) {
 	serverSupported := options.ServerConfigurationControl != nil
 	modeOptions := []string{sessionModeClient}
@@ -303,15 +268,16 @@ func (m configuratorSessionModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		m.adjustInputsToViewport()
 		if m.tab == configuratorTabLogs {
 			m.logs.ensure(m.width, m.height, m.preferences, "", configuratorLogsHint)
-			m.logs.refresh(m.logsFeed(), m.preferences)
+			m.logs.refresh(globalRuntimeLogFeed(), m.preferences)
 		}
 		return m, nil
 	case logViewportTickMsg:
 		if msg.seq != m.logs.tickSeq || m.tab != configuratorTabLogs {
 			return m, nil
 		}
-		m.logs.refresh(m.logsFeed(), m.preferences)
-		return m, configuratorLogUpdateCmd(m.logsFeed(), m.logs.waitStop, m.logs.tickSeq)
+		feed := globalRuntimeLogFeed()
+		m.logs.refresh(feed, m.preferences)
+		return m, logViewportUpdateCmd(feed, m.logs.waitStop, m.logs.tickSeq)
 	case pasteSettledMsg:
 		if m.screen == configuratorScreenClientAddJSON && msg.seq == m.client.pasteSeq {
 			m.tryFormatJSON()
@@ -398,172 +364,6 @@ func (m configuratorSessionModel) View() tea.View {
 	return v
 }
 
-func (m configuratorSessionModel) mainTabView() string {
-	switch m.screen {
-	case configuratorScreenMode:
-		return m.renderSelectionScreen(
-			"Select mode",
-			m.notice,
-			m.modeOptions,
-			m.cursor,
-			"up/k down/j move | Enter select | Tab switch tabs | Esc exit | ctrl+c exit",
-		)
-	case configuratorScreenClientSelect:
-		clientSelectHint := "up/k down/j move | Enter select | Tab switch tabs | Esc back | ctrl+c exit"
-		if len(m.modeOptions) == 1 {
-			clientSelectHint = "up/k down/j move | Enter select | Tab switch tabs | Esc exit | ctrl+c exit"
-		}
-		return m.renderSelectionScreen(
-			"Select configuration - or add/remove one:",
-			m.notice,
-			m.client.menuOptions,
-			m.cursor,
-			clientSelectHint,
-		)
-	case configuratorScreenClientRemove:
-		return m.renderSelectionScreen(
-			"Choose a configuration to remove:",
-			"",
-			m.client.removePaths,
-			m.cursor,
-			"up/k down/j move | Enter remove | Tab switch tabs | Esc back | ctrl+c exit",
-		)
-	case configuratorScreenClientAddName:
-		styles := resolveUIStyles(m.preferences)
-		container := styles.inputFrame.Width(m.inputContainerWidth())
-		stats := styles.meta.Render("Characters: " + formatCount(utf8.RuneCountInString(m.client.addNameInput.Value()), m.client.addNameInput.CharLimit))
-		body := make([]string, 0, 4)
-		if strings.TrimSpace(m.notice) != "" {
-			body = append(body, m.notice, "")
-		}
-		body = append(body, container.Render(m.client.addNameInput.View()), stats)
-		return renderScreen(
-			m.width,
-			m.height,
-			m.tabsLine(styles),
-			"Name configuration",
-			body,
-			"Enter confirm | Tab switch tabs | Esc back | ctrl+c exit",
-			m.preferences,
-			styles,
-		)
-	case configuratorScreenClientAddJSON:
-		styles := resolveUIStyles(m.preferences)
-		container := styles.inputFrame.Width(m.inputContainerWidth())
-		lines := 1
-		if value := m.client.addJSONInput.Value(); value != "" {
-			lines = len(strings.Split(value, "\n"))
-		}
-		stats := styles.meta.Render(fmt.Sprintf("Lines: %d", lines))
-		body := make([]string, 0, 4)
-		if strings.TrimSpace(m.notice) != "" {
-			body = append(body, m.notice, "")
-		}
-		body = append(body, container.Render(m.client.addJSONInput.View()), stats)
-		return renderScreen(
-			m.width,
-			m.height,
-			m.tabsLine(styles),
-			"Paste configuration",
-			body,
-			"Enter confirm | Esc back | ctrl+c exit",
-			m.preferences,
-			styles,
-		)
-	case configuratorScreenClientInvalid:
-		options := []string{sessionInvalidOK}
-		if m.client.invalidAllowDelete {
-			options = []string{sessionInvalidDelete, sessionInvalidOK}
-		}
-		subtitle := "Configuration is invalid: " + summarizeInvalidConfigurationError(m.client.invalidErr)
-		return m.renderSelectionScreen(
-			"Configuration error",
-			subtitle,
-			options,
-			m.cursor,
-			"up/k down/j move | Enter select | Tab switch tabs | Esc back | ctrl+c exit",
-		)
-	case configuratorScreenServerSelect:
-		return m.renderSelectionScreen(
-			"Choose an option",
-			m.notice,
-			m.server.menuOptions,
-			m.cursor,
-			"up/k down/j move | Enter select | Tab switch tabs | Esc back | ctrl+c exit",
-		)
-	case configuratorScreenServerManage:
-		return m.renderSelectionScreen(
-			"Select client to enable/disable or delete",
-			"",
-			m.server.manageLabels,
-			m.cursor,
-			"up/k down/j move | Enter toggle | d delete | Tab switch tabs | Esc back | ctrl+c exit",
-		)
-	case configuratorScreenServerDeleteConfirm:
-		return m.renderSelectionScreen(
-			fmt.Sprintf(
-				"Delete client #%d %s?",
-				m.server.deletePeer.ClientID,
-				serverPeerDisplayName(m.server.deletePeer),
-			),
-			"This action removes client access from server configuration.",
-			[]string{sessionServerDeleteConfirm, sessionCancel},
-			m.cursor,
-			"up/k down/j move | Enter confirm | Tab switch tabs | Esc back | ctrl+c exit",
-		)
-	case configuratorScreenDaemonManage:
-		return m.renderDaemonManageScreen()
-	case configuratorScreenDaemonReconfigureConfirm:
-		roleLabel := "selected role"
-		switch m.pendingDaemonMode {
-		case runtime.ModeClient:
-			roleLabel = "client"
-		case runtime.ModeServer:
-			roleLabel = "server"
-		}
-		return m.renderSelectionScreen(
-			"Daemon is active",
-			fmt.Sprintf("Applying %s daemon setup requires restart. Continue now?", roleLabel),
-			[]string{sessionDaemonConfirmReconfigureNow, sessionCancel},
-			m.cursor,
-			"up/k down/j move | Enter select | Tab switch tabs | Esc back | ctrl+c exit",
-		)
-	case configuratorScreenDaemonActiveConfirm:
-		modeLabel := "selected mode"
-		switch m.pendingStartMode {
-		case runtime.ModeClient:
-			modeLabel = "client"
-		case runtime.ModeServer:
-			modeLabel = "server"
-		}
-		notice := fmt.Sprintf("tungo.service is active. Stop it before starting %s in TUI mode.", modeLabel)
-		if strings.TrimSpace(m.notice) != "" {
-			notice = m.notice + "\n" + notice
-		}
-		return m.renderSelectionScreen(
-			"Active daemon detected",
-			notice,
-			[]string{sessionStopDaemonContinue, sessionCancel},
-			m.cursor,
-			"up/k down/j move | Enter select | Tab switch tabs | Esc back | ctrl+c exit",
-		)
-	case configuratorScreenDaemonCheckErrorConfirm:
-		subtitle := "Failed to check daemon status."
-		if strings.TrimSpace(m.notice) != "" {
-			subtitle = m.notice
-		}
-		return m.renderSelectionScreen(
-			"Cannot verify daemon status",
-			subtitle,
-			[]string{sessionRetryDaemonCheck, sessionStartAnywayUnsafe, sessionCancel},
-			m.cursor,
-			"up/k down/j move | Enter select | Tab switch tabs | Esc back | ctrl+c exit",
-		)
-	default:
-		return ""
-	}
-}
-
 func (m configuratorSessionModel) updateModeScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
 	switch msg.String() {
 	case "esc":
@@ -600,736 +400,6 @@ func (m configuratorSessionModel) updateModeScreen(msg tea.KeyPressMsg) (tea.Mod
 	return m, nil
 }
 
-func (m configuratorSessionModel) updateClientSelectScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		m.notice = ""
-		m.cursor = 0
-		if len(m.modeOptions) == 1 {
-			m.resultErr = ErrConfiguratorSessionUserExit
-			m.done = true
-			return m, tea.Quit
-		}
-		m.screen = configuratorScreenMode
-		return m, nil
-	}
-
-	m.updateCursor(msg, len(m.client.menuOptions))
-	if msg.String() != "enter" || len(m.client.menuOptions) == 0 {
-		return m, nil
-	}
-
-	selected := m.client.menuOptions[m.cursor]
-	switch selected {
-	case sessionClientAdd:
-		m.notice = ""
-		m.cursor = 0
-		m.screen = configuratorScreenClientAddName
-		m.initNameInput()
-		m.adjustInputsToViewport()
-		return m, textinput.Blink
-	case sessionClientRemove:
-		if len(m.client.configs) == 0 {
-			m.notice = "No configurations available for removal."
-			return m, nil
-		}
-		m.notice = ""
-		m.cursor = 0
-		m.screen = configuratorScreenClientRemove
-		m.client.removePaths = append([]string(nil), m.client.configs...)
-		return m, nil
-	default:
-		if err := m.options.ClientConfigurationControl.Select(selected); err != nil {
-			m.resultErr = err
-			m.done = true
-			return m, tea.Quit
-		}
-
-		cfgErr := m.options.ClientConfigurationControl.ValidateActive()
-		if isInvalidClientConfigurationError(cfgErr) {
-			m.client.invalidErr = cfgErr
-			m.client.invalidConfig = selected
-			m.client.invalidAllowDelete = true
-			m.cursor = 0
-			m.screen = configuratorScreenClientInvalid
-			return m, nil
-		}
-		if cfgErr != nil {
-			m.resultErr = cfgErr
-			m.done = true
-			return m, tea.Quit
-		}
-
-		m = m.startModeWithDaemonGuard(runtime.ModeClient, configuratorScreenClientSelect, false)
-		if m.done {
-			m = m.persistAutoSelectClientConfig(selected)
-			return m, tea.Quit
-		}
-		if isDaemonStartConfirmationScreen(m.screen) {
-			m.pendingClientConfig = selected
-		}
-		return m, nil
-	}
-}
-
-func (m configuratorSessionModel) updateClientRemoveScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		m.notice = ""
-		m.cursor = 0
-		m.screen = configuratorScreenClientSelect
-		return m, nil
-	}
-
-	m.updateCursor(msg, len(m.client.removePaths))
-	if msg.String() != "enter" || len(m.client.removePaths) == 0 {
-		return m, nil
-	}
-
-	toDelete := m.client.removePaths[m.cursor]
-	if err := m.options.ClientConfigurationControl.Delete(toDelete); err != nil {
-		m.resultErr = err
-		m.done = true
-		return m, tea.Quit
-	}
-	if err := m.reloadClientConfigs(); err != nil {
-		m.resultErr = err
-		m.done = true
-		return m, tea.Quit
-	}
-	m.notice = "Configuration removed."
-	m.cursor = 0
-	m.screen = configuratorScreenClientSelect
-	return m, nil
-}
-
-func (m configuratorSessionModel) updateClientAddNameScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		m.notice = ""
-		m.cursor = 0
-		m.screen = configuratorScreenClientSelect
-		return m, nil
-	case "enter":
-		name := strings.TrimSpace(m.client.addNameInput.Value())
-		if name == "" {
-			m.notice = "Configuration name cannot be empty."
-			return m, nil
-		}
-		m.client.addName = name
-		m.notice = ""
-		m.cursor = 0
-		m.screen = configuratorScreenClientAddJSON
-		m.client.lastInputAt = time.Time{}
-		m.initJSONInput()
-		m.adjustInputsToViewport()
-		return m, textarea.Blink
-	}
-
-	var cmd tea.Cmd
-	m.client.addNameInput, cmd = m.client.addNameInput.Update(msg)
-	return m, cmd
-}
-
-const pasteDebounce = 300 * time.Millisecond
-
-func (m configuratorSessionModel) updateClientAddJSONScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
-	if msg.String() == "esc" {
-		m.notice = ""
-		m.screen = configuratorScreenClientAddName
-		m.adjustInputsToViewport()
-		return m, nil
-	}
-
-	if msg.String() == "enter" {
-		// Debounce: if Enter arrives within pasteDebounce of the last
-		// non-Enter keystroke, it is almost certainly a newline from a
-		// character-by-character terminal paste — insert it as a newline
-		// instead of submitting.
-		if !m.client.lastInputAt.IsZero() && time.Since(m.client.lastInputAt) < pasteDebounce {
-			m.client.lastInputAt = time.Now()
-			var cmd tea.Cmd
-			m.client.addJSONInput, cmd = m.client.addJSONInput.Update(msg)
-			return m, cmd
-		}
-
-		if err := m.options.ClientConfigurationControl.CreateFromJSON(m.client.addName, m.client.addJSONInput.Value()); err != nil {
-			if isInvalidClientConfigurationError(err) {
-				m.client.invalidErr = err
-				m.client.invalidConfig = ""
-				m.client.invalidAllowDelete = false
-				m.cursor = 0
-				m.screen = configuratorScreenClientInvalid
-				return m, nil
-			}
-			m.resultErr = err
-			m.done = true
-			return m, tea.Quit
-		}
-		if err := m.reloadClientConfigs(); err != nil {
-			m.resultErr = err
-			m.done = true
-			return m, tea.Quit
-		}
-
-		m.notice = "Configuration added."
-		m.cursor = 0
-		m.screen = configuratorScreenClientSelect
-		return m, nil
-	}
-
-	// Track non-Enter input timing for debounce.
-	m.client.lastInputAt = time.Now()
-	m.client.pasteSeq++
-	seq := m.client.pasteSeq
-
-	// Forward to textarea (paste characters, cursor movement, etc.)
-	var cmd tea.Cmd
-	m.client.addJSONInput, cmd = m.client.addJSONInput.Update(msg)
-	return m, tea.Batch(cmd, tea.Tick(pasteDebounce, func(time.Time) tea.Msg {
-		return pasteSettledMsg{seq: seq}
-	}))
-}
-
-func (m configuratorSessionModel) updateClientInvalidScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		m.notice = ""
-		m.cursor = 0
-		m.screen = configuratorScreenClientSelect
-		return m, nil
-	}
-
-	options := []string{sessionInvalidOK}
-	if m.client.invalidAllowDelete {
-		options = []string{sessionInvalidDelete, sessionInvalidOK}
-	}
-	m.updateCursor(msg, len(options))
-	if msg.String() != "enter" || len(options) == 0 {
-		return m, nil
-	}
-
-	selected := options[m.cursor]
-	if selected == sessionInvalidDelete && m.client.invalidAllowDelete {
-		if strings.TrimSpace(m.client.invalidConfig) == "" {
-			m.resultErr = errors.New("invalid configuration cannot be deleted")
-			m.done = true
-			return m, tea.Quit
-		}
-		if err := m.options.ClientConfigurationControl.Delete(m.client.invalidConfig); err != nil {
-			m.resultErr = err
-			m.done = true
-			return m, tea.Quit
-		}
-		if err := m.reloadClientConfigs(); err != nil {
-			m.resultErr = err
-			m.done = true
-			return m, tea.Quit
-		}
-		m.notice = "Invalid configuration deleted."
-	}
-	m.cursor = 0
-	m.screen = configuratorScreenClientSelect
-	return m, nil
-}
-
-func (m configuratorSessionModel) updateServerSelectScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		m.notice = ""
-		m.cursor = 0
-		m.screen = configuratorScreenMode
-		return m, nil
-	}
-
-	m.updateCursor(msg, len(m.server.menuOptions))
-	if msg.String() != "enter" || len(m.server.menuOptions) == 0 {
-		return m, nil
-	}
-
-	switch m.server.menuOptions[m.cursor] {
-	case sessionServerStart:
-		m = m.startModeWithDaemonGuard(runtime.ModeServer, configuratorScreenServerSelect, false)
-		if m.done {
-			return m, tea.Quit
-		}
-		return m, nil
-	case sessionServerAdd:
-		generated, err := m.options.ServerConfigurationControl.GenerateClientConfiguration()
-		if err != nil {
-			m.resultErr = err
-			m.done = true
-			return m, tea.Quit
-		}
-		m.notice = fmt.Sprintf("Client configuration saved to %s", generated.Path)
-		return m, nil
-	case sessionServerManage:
-		peers, err := m.options.ServerConfigurationControl.ListPeers()
-		if err != nil {
-			m.resultErr = err
-			m.done = true
-			return m, tea.Quit
-		}
-		if len(peers) == 0 {
-			m.notice = "No clients configured yet."
-			return m, nil
-		}
-		m.server.managePeers = peers
-		m.server.manageLabels = buildServerManageLabels(peers)
-		m.notice = ""
-		m.cursor = 0
-		m.screen = configuratorScreenServerManage
-		return m, nil
-	}
-	return m, nil
-}
-
-func (m configuratorSessionModel) updateServerManageScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		m.notice = ""
-		m.cursor = 0
-		m.screen = configuratorScreenServerSelect
-		return m, nil
-	case "d", "D":
-		if len(m.server.managePeers) == 0 {
-			return m, nil
-		}
-		m.server.deletePeer = m.server.managePeers[m.cursor]
-		m.server.deleteCursor = m.cursor
-		m.cursor = 0
-		m.screen = configuratorScreenServerDeleteConfirm
-		return m, nil
-	}
-
-	m.updateCursor(msg, len(m.server.managePeers))
-	if msg.String() != "enter" || len(m.server.managePeers) == 0 {
-		return m, nil
-	}
-
-	peer := m.server.managePeers[m.cursor]
-	nextEnabled := !peer.Enabled
-	if err := m.options.ServerConfigurationControl.SetPeerEnabled(peer.ClientID, nextEnabled); err != nil {
-		m.notice = fmt.Sprintf("Failed to update client #%d: %v", peer.ClientID, err)
-		m.screen = configuratorScreenServerSelect
-		m.cursor = 0
-		return m, nil
-	}
-
-	peers, err := m.options.ServerConfigurationControl.ListPeers()
-	if err != nil {
-		m.resultErr = err
-		m.done = true
-		return m, tea.Quit
-	}
-	if len(peers) == 0 {
-		m.notice = "No clients configured yet."
-		m.screen = configuratorScreenServerSelect
-		m.cursor = 0
-		return m, nil
-	}
-
-	m.server.managePeers = peers
-	m.server.manageLabels = buildServerManageLabels(peers)
-	if m.cursor >= len(m.server.managePeers) {
-		m.cursor = len(m.server.managePeers) - 1
-	}
-	return m, nil
-}
-
-func (m configuratorSessionModel) updateServerDeleteConfirmScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		if len(m.server.managePeers) > 0 {
-			m.cursor = minInt(m.server.deleteCursor, len(m.server.managePeers)-1)
-		} else {
-			m.cursor = 0
-		}
-		m.screen = configuratorScreenServerManage
-		return m, nil
-	}
-
-	options := []string{sessionServerDeleteConfirm, sessionCancel}
-	m.updateCursor(msg, len(options))
-	if msg.String() != "enter" {
-		return m, nil
-	}
-
-	selected := options[m.cursor]
-	if selected == sessionCancel {
-		if len(m.server.managePeers) > 0 {
-			m.cursor = minInt(m.server.deleteCursor, len(m.server.managePeers)-1)
-		} else {
-			m.cursor = 0
-		}
-		m.screen = configuratorScreenServerManage
-		return m, nil
-	}
-
-	if err := m.options.ServerConfigurationControl.RemovePeer(m.server.deletePeer.ClientID); err != nil {
-		m.notice = fmt.Sprintf("Failed to remove client #%d: %v", m.server.deletePeer.ClientID, err)
-		m.screen = configuratorScreenServerManage
-		m.cursor = 0
-		return m, nil
-	}
-
-	peers, err := m.options.ServerConfigurationControl.ListPeers()
-	if err != nil {
-		m.resultErr = err
-		m.done = true
-		return m, tea.Quit
-	}
-	if len(peers) == 0 {
-		m.notice = "No clients configured yet."
-		m.screen = configuratorScreenServerSelect
-		m.cursor = 0
-		return m, nil
-	}
-
-	m.notice = fmt.Sprintf(
-		"Client #%d %s removed.",
-		m.server.deletePeer.ClientID,
-		serverPeerDisplayName(m.server.deletePeer),
-	)
-	m.server.managePeers = peers
-	m.server.manageLabels = buildServerManageLabels(peers)
-	m.cursor = minInt(m.server.deleteCursor, len(peers)-1)
-	m.screen = configuratorScreenServerManage
-	return m, nil
-}
-
-func (m configuratorSessionModel) updateDaemonManageScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		return m.leaveDaemonManageScreen(), nil
-	}
-
-	m.updateCursor(msg, len(m.daemon.menuOptions))
-	if msg.String() != "enter" || len(m.daemon.menuOptions) == 0 {
-		return m, nil
-	}
-
-	selected := m.daemon.menuOptions[m.cursor]
-	selectedCursor := m.cursor
-	var err error
-	switch selected {
-	case sessionDaemonSetupClient:
-		m, err = m.applyDaemonSetup(runtime.ModeClient, false)
-		if err != nil {
-			m.notice = err.Error()
-			return m, nil
-		}
-	case sessionDaemonSetupServer:
-		m, err = m.applyDaemonSetup(runtime.ModeServer, false)
-		if err != nil {
-			m.notice = err.Error()
-			return m, nil
-		}
-	case sessionDaemonReconfClient:
-		if daemonStateBlocksRuntimeStart(string(m.daemon.status.ActiveState)) {
-			m.pendingDaemonMode = runtime.ModeClient
-			m.cursor = 0
-			m.screen = configuratorScreenDaemonReconfigureConfirm
-			return m, nil
-		}
-		m, err = m.applyDaemonSetup(runtime.ModeClient, false)
-		if err != nil {
-			m.notice = err.Error()
-			return m, nil
-		}
-	case sessionDaemonReconfServer:
-		if daemonStateBlocksRuntimeStart(string(m.daemon.status.ActiveState)) {
-			m.pendingDaemonMode = runtime.ModeServer
-			m.cursor = 0
-			m.screen = configuratorScreenDaemonReconfigureConfirm
-			return m, nil
-		}
-		m, err = m.applyDaemonSetup(runtime.ModeServer, false)
-		if err != nil {
-			m.notice = err.Error()
-			return m, nil
-		}
-	case sessionDaemonStart:
-		if m.options.Daemon == nil {
-			m.notice = "Daemon start is unavailable."
-			return m, nil
-		}
-		if err := m.options.Daemon.StartUnit(); err != nil {
-			m.notice = fmt.Sprintf("Failed to start daemon: %v", err)
-			return m, nil
-		}
-		m.notice = ""
-	case sessionDaemonStop:
-		if m.options.Daemon == nil {
-			m.notice = "Daemon stop is unavailable."
-			return m, nil
-		}
-		if err := m.options.Daemon.StopUnit(); err != nil {
-			m.notice = fmt.Sprintf("Failed to stop daemon: %v", err)
-			return m, nil
-		}
-		m.notice = ""
-	case sessionDaemonEnable:
-		if m.options.Daemon == nil {
-			m.notice = "Daemon enable is unavailable."
-			return m, nil
-		}
-		if err := m.options.Daemon.EnableUnit(); err != nil {
-			m.notice = fmt.Sprintf("Failed to enable daemon: %v", err)
-			return m, nil
-		}
-		m.notice = ""
-	case sessionDaemonDisable:
-		if m.options.Daemon == nil {
-			m.notice = "Daemon disable is unavailable."
-			return m, nil
-		}
-		if err := m.options.Daemon.DisableUnit(); err != nil {
-			m.notice = fmt.Sprintf("Failed to disable daemon: %v", err)
-			return m, nil
-		}
-		m.notice = ""
-	case sessionDaemonDelete:
-		if m.options.Daemon == nil {
-			m.notice = "Daemon remove is unavailable."
-			return m, nil
-		}
-		if err := m.options.Daemon.RemoveUnit(); err != nil {
-			m.notice = fmt.Sprintf("Failed to remove daemon: %v", err)
-			return m, nil
-		}
-		m.notice = ""
-	default:
-		return m, nil
-	}
-
-	m.refreshDaemonStatus()
-	m.cursor = daemonMenuCursorAfterRefresh(m.daemon.menuOptions, selected, selectedCursor)
-	return m, nil
-}
-
-func (m configuratorSessionModel) updateDaemonReconfigureConfirmScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		m.screen = configuratorScreenDaemonManage
-		m.cursor = 0
-		m.pendingDaemonMode = 0
-		m.notice = "Reconfigure cancelled."
-		return m, nil
-	}
-
-	options := []string{sessionDaemonConfirmReconfigureNow, sessionCancel}
-	m.updateCursor(msg, len(options))
-	if msg.String() != "enter" {
-		return m, nil
-	}
-
-	if options[m.cursor] == sessionCancel {
-		m.screen = configuratorScreenDaemonManage
-		m.cursor = 0
-		m.pendingDaemonMode = 0
-		m.notice = "Reconfigure cancelled."
-		return m, nil
-	}
-
-	targetMode := m.pendingDaemonMode
-	m.pendingDaemonMode = 0
-	m.screen = configuratorScreenDaemonManage
-	m.cursor = 0
-
-	updated, err := m.applyDaemonSetup(targetMode, true)
-	if err != nil {
-		updated.notice = err.Error()
-		return updated, nil
-	}
-	return updated, nil
-}
-
-func (m configuratorSessionModel) applyDaemonSetup(targetMode runtime.Mode, restartRunning bool) (configuratorSessionModel, error) {
-	if m.options.Daemon == nil {
-		return m, errors.New("daemon setup is unavailable")
-	}
-	if targetMode == runtime.ModeClient {
-		if err := m.options.ClientConfigurationControl.ValidateActive(); err != nil {
-			return m, fmt.Errorf("cannot setup client daemon: %v", err)
-		}
-	}
-
-	path, err := m.options.Daemon.Setup(targetMode)
-	if err != nil {
-		return m, fmt.Errorf("failed to setup daemon: %v", err)
-	}
-
-	if restartRunning {
-		role := "Client"
-		if targetMode == runtime.ModeServer {
-			role = "Server"
-		}
-		m.notice = fmt.Sprintf("%s daemon reconfigured at %s and restarted.", role, path)
-	} else {
-		m.notice = ""
-	}
-	m.refreshDaemonStatus()
-	m.cursor = 0
-	return m, nil
-}
-
-func (m configuratorSessionModel) updateDaemonActiveConfirmScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		m = m.cancelPendingDaemonStart("Start cancelled.")
-		return m, nil
-	}
-
-	options := []string{sessionStopDaemonContinue, sessionCancel}
-	m.updateCursor(msg, len(options))
-	if msg.String() != "enter" {
-		return m, nil
-	}
-
-	selected := options[m.cursor]
-	if selected == sessionCancel {
-		m = m.cancelPendingDaemonStart("Start cancelled.")
-		return m, nil
-	}
-
-	if m.options.Daemon == nil {
-		m = m.cancelPendingDaemonStart("Stopping daemon is unavailable.")
-		return m, nil
-	}
-
-	if err := m.options.Daemon.StopUnit(); err != nil {
-		m = m.cancelPendingDaemonStart(fmt.Sprintf("Failed to stop daemon: %v", err))
-		return m, nil
-	}
-
-	return m.completePendingDaemonStart("Daemon stopped. Starting selected mode.")
-}
-
-func (m configuratorSessionModel) updateDaemonCheckErrorConfirmScreen(msg tea.KeyPressMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "esc":
-		m = m.cancelPendingDaemonStart("Start cancelled.")
-		return m, nil
-	}
-
-	options := []string{sessionRetryDaemonCheck, sessionStartAnywayUnsafe, sessionCancel}
-	m.updateCursor(msg, len(options))
-	if msg.String() != "enter" {
-		return m, nil
-	}
-
-	selected := options[m.cursor]
-	switch selected {
-	case sessionCancel:
-		m = m.cancelPendingDaemonStart("Start cancelled.")
-		return m, nil
-	case sessionStartAnywayUnsafe:
-		return m.completePendingDaemonStart("Daemon status check failed. Starting selected mode without daemon guard.")
-	case sessionRetryDaemonCheck:
-		targetMode := m.pendingStartMode
-		returnScreen := m.pendingStartScreen
-		pendingClientConfig := m.pendingClientConfig
-		m = m.startModeWithDaemonGuard(targetMode, returnScreen, true)
-		if m.done {
-			return m, tea.Quit
-		}
-		if !m.done && isDaemonStartConfirmationScreen(m.screen) && targetMode == runtime.ModeClient {
-			m.pendingClientConfig = pendingClientConfig
-		}
-		return m, nil
-	default:
-		return m, nil
-	}
-}
-
-func (m configuratorSessionModel) completePendingDaemonStart(notice string) (configuratorSessionModel, tea.Cmd) {
-	targetMode := m.pendingStartMode
-	pendingClientConfig := m.pendingClientConfig
-	m = m.clearPendingDaemonStart()
-	if targetMode == runtime.ModeClient {
-		m = m.persistAutoSelectClientConfig(pendingClientConfig)
-	}
-	m.notice = notice
-	m.resultMode = targetMode
-	m.done = true
-	return m, tea.Quit
-}
-
-func (m configuratorSessionModel) startModeWithDaemonGuard(targetMode runtime.Mode, returnScreen configuratorScreen, preserveNotice bool) configuratorSessionModel {
-	m = m.clearPendingDaemonStart()
-
-	if m.options.Daemon == nil {
-		m.resultMode = targetMode
-		m.done = true
-		return m
-	}
-
-	active, err := m.options.Daemon.IsUnitActive()
-	if err != nil {
-		message := fmt.Sprintf("Failed to check daemon status: %v", err)
-		if preserveNotice {
-			m.notice = appendNotice(m.notice, message)
-		} else {
-			m.notice = message
-		}
-		m.cursor = 0
-		m.pendingStartMode = targetMode
-		m.pendingStartScreen = returnScreen
-		m.screen = configuratorScreenDaemonCheckErrorConfirm
-		return m
-	}
-	if !active {
-		m.resultMode = targetMode
-		m.done = true
-		return m
-	}
-	if !preserveNotice {
-		m.notice = ""
-	}
-	m.cursor = 0
-	m.pendingStartMode = targetMode
-	m.pendingStartScreen = returnScreen
-	m.screen = configuratorScreenDaemonActiveConfirm
-	return m
-}
-
-func (m configuratorSessionModel) cancelPendingDaemonStart(notice string) configuratorSessionModel {
-	returnScreen := m.pendingStartScreen
-	m = m.clearPendingDaemonStart()
-	m.notice = notice
-	m.cursor = 0
-	m.screen = returnScreen
-	return m
-}
-
-func (m configuratorSessionModel) clearPendingDaemonStart() configuratorSessionModel {
-	m.pendingStartMode = 0
-	m.pendingStartScreen = configuratorScreenMode
-	m.pendingClientConfig = ""
-	return m
-}
-
-func isDaemonStartConfirmationScreen(screen configuratorScreen) bool {
-	switch screen {
-	case configuratorScreenDaemonActiveConfirm, configuratorScreenDaemonCheckErrorConfirm:
-		return true
-	default:
-		return false
-	}
-}
-
-func (m configuratorSessionModel) persistAutoSelectClientConfig(selected string) configuratorSessionModel {
-	if strings.TrimSpace(selected) == "" {
-		return m
-	}
-	p := m.settings.Preferences()
-	p.AutoSelectClientConfig = selected
-	m.settings.update(p)
-	_ = savePreferencesToDisk(p)
-	return m
-}
-
 func appendNotice(existing, next string) string {
 	next = strings.TrimSpace(next)
 	if next == "" {
@@ -1357,8 +427,9 @@ func (m configuratorSessionModel) cycleTab() (tea.Model, tea.Cmd) {
 		m.logs.restartWait()
 		m.logs.tickSeq++
 		m.logs.ensure(m.width, m.height, m.preferences, "", configuratorLogsHint)
-		m.logs.refresh(m.logsFeed(), m.preferences)
-		return m, configuratorLogUpdateCmd(m.logsFeed(), m.logs.waitStop, m.logs.tickSeq)
+		feed := globalRuntimeLogFeed()
+		m.logs.refresh(feed, m.preferences)
+		return m, logViewportUpdateCmd(feed, m.logs.waitStop, m.logs.tickSeq)
 	}
 	if previous == configuratorTabLogs {
 		m.logs.stopWait()
@@ -1411,399 +482,8 @@ func (m configuratorSessionModel) updateLogsTab(msg tea.KeyPressMsg) (tea.Model,
 	return m, m.logs.updateKeys(msg, defaultSelectorKeyMap())
 }
 
-func (m *configuratorSessionModel) refreshDaemonStatus() {
-	if m.options.Daemon == nil {
-		m.daemon.statusErr = errors.New("daemon management is unavailable")
-		m.daemon.status = systemd.UnitStatus{}
-		m.daemon.menuOptions = nil
-		m.daemon.updatedAt = time.Time{}
-		return
-	}
-
-	status, err := m.options.Daemon.Status()
-	if err != nil {
-		m.daemon.statusErr = err
-		m.daemon.status = systemd.UnitStatus{}
-		m.daemon.menuOptions = nil
-		m.daemon.updatedAt = time.Time{}
-		return
-	}
-	m.daemon.statusErr = nil
-	m.daemon.status = status
-	m.daemon.menuOptions = m.daemonMenuOptions(status)
-	m.daemon.updatedAt = time.Now()
-}
-
-func (m configuratorSessionModel) daemonMenuOptions(status systemd.UnitStatus) []string {
-	if m.options.Daemon == nil {
-		return nil
-	}
-	options := make([]string, 0, 7)
-	if !status.Installed {
-		options = append(options, sessionDaemonSetupClient)
-		if m.serverSupported {
-			options = append(options, sessionDaemonSetupServer)
-		}
-		return options
-	}
-
-	activeBlocksStart := daemonStateBlocksRuntimeStart(string(status.ActiveState))
-	if activeBlocksStart {
-		options = append(options, sessionDaemonStop)
-	}
-	if !activeBlocksStart && daemonStateAllowsStart(string(status.ActiveState)) {
-		options = append(options, sessionDaemonStart)
-	}
-	if daemonUnitFileStateIsEnabled(string(status.UnitFileState)) {
-		options = append(options, sessionDaemonDisable)
-	}
-	if daemonUnitFileStateIsDisabled(string(status.UnitFileState)) {
-		options = append(options, sessionDaemonEnable)
-	}
-	options = append(options, sessionDaemonReconfClient)
-	if m.serverSupported {
-		options = append(options, sessionDaemonReconfServer)
-	}
-	if status.Managed {
-		options = append(options, sessionDaemonDelete)
-	}
-	return options
-}
-
-func (m configuratorSessionModel) daemonNotice() string {
-	statusLine := m.daemonStatusLine()
-	notice := strings.TrimSpace(m.notice)
-	if notice == "" {
-		return statusLine
-	}
-	return statusLine + "\n" + notice
-}
-
-func (m configuratorSessionModel) daemonStatusLine() string {
-	if m.daemon.statusErr != nil {
-		return "Status error: " + m.daemon.statusErr.Error()
-	}
-	loadState := normalizeDaemonStateField(string(m.daemon.status.LoadState))
-	unitFileState := normalizeDaemonStateField(string(m.daemon.status.UnitFileState))
-	activeState := normalizeDaemonStateField(string(m.daemon.status.ActiveState))
-	subState := normalizeDaemonStateField(m.daemon.status.SubState)
-	result := normalizeDaemonStateField(m.daemon.status.Result)
-	execMainStatus := normalizeDaemonStateField(m.daemon.status.ExecMainStatus)
-	execStart := normalizeDaemonRawField(m.daemon.status.ExecStart)
-	fragmentPath := normalizeDaemonRawField(m.daemon.status.FragmentPath)
-	derivedRole, derivedRoleSource := daemonDerivedRole(m.daemon.status, execStart)
-	return strings.Join([]string{
-		fmt.Sprintf("Active: %s", activeState),
-		fmt.Sprintf("Sub: %s", subState),
-		fmt.Sprintf("Result: %s", result),
-		fmt.Sprintf("UnitFile: %s", unitFileState),
-		fmt.Sprintf("Load: %s", loadState),
-		fmt.Sprintf("ExecMainStatus: %s", execMainStatus),
-		fmt.Sprintf("ExecStart: %s", execStart),
-		fmt.Sprintf("FragmentPath: %s", fragmentPath),
-		fmt.Sprintf("DerivedRole: %s (from %s)", derivedRole, derivedRoleSource),
-	}, "\n")
-}
-
-func normalizeDaemonStateField(value string) string {
-	normalized := strings.ToLower(strings.TrimSpace(value))
-	if normalized == "" {
-		return "unknown"
-	}
-	return normalized
-}
-
-func normalizeDaemonRawField(value string) string {
-	normalized := strings.TrimSpace(value)
-	if normalized == "" {
-		return "unknown"
-	}
-	return normalized
-}
-
-func daemonRoleFromExecStart(execStart string) string {
-	raw := strings.ToLower(strings.TrimSpace(execStart))
-	if raw == "" || raw == "unknown" {
-		return "unknown"
-	}
-	if strings.Contains(raw, " tungo c") || strings.Contains(raw, "/tungo c") {
-		return "client"
-	}
-	if strings.Contains(raw, " tungo s") || strings.Contains(raw, "/tungo s") {
-		return "server"
-	}
-	return "unknown"
-}
-
-func daemonDerivedRole(status systemd.UnitStatus, execStart string) (string, string) {
-	if role := daemonRoleFromExecStart(execStart); role != "unknown" {
-		return role, "ExecStart"
-	}
-	switch status.Role {
-	case systemd.UnitRoleClient:
-		return "client", "Role"
-	case systemd.UnitRoleServer:
-		return "server", "Role"
-	default:
-		return "unknown", "Role"
-	}
-}
-
-func daemonStateBlocksRuntimeStart(activeState string) bool {
-	switch normalizeDaemonStateField(activeState) {
-	case "active", "reloading", "activating", "deactivating":
-		return true
-	default:
-		return false
-	}
-}
-
-func daemonStateAllowsStart(activeState string) bool {
-	switch normalizeDaemonStateField(activeState) {
-	case "inactive", "failed":
-		return true
-	default:
-		return false
-	}
-}
-
-func daemonUnitFileStateIsEnabled(unitFileState string) bool {
-	return normalizeDaemonStateField(unitFileState) == "enabled"
-}
-
-func daemonUnitFileStateIsDisabled(unitFileState string) bool {
-	return normalizeDaemonStateField(unitFileState) == "disabled"
-}
-
-func (m configuratorSessionModel) leaveDaemonManageScreen() configuratorSessionModel {
-	m.tab = configuratorTabMain
-	m.screen = configuratorScreenMode
-	if idx := slices.Index(m.modeOptions, sessionModeDaemon); idx >= 0 {
-		m.cursor = idx
-	} else {
-		m.cursor = 0
-	}
-	m.pendingDaemonMode = 0
-	m.refreshDaemonStatus()
-	return m
-}
-
 func (m configuratorSessionModel) settingsRows() []string {
 	return uiSettingsRows(m.preferences, m.serverSupported)
-}
-
-func (m *configuratorSessionModel) reloadClientConfigs() error {
-	configs, err := m.options.ClientConfigurationControl.List()
-	if err != nil {
-		return err
-	}
-	m.client.configs = configs
-	m.client.menuOptions = make([]string, 0, len(configs)+3)
-	m.client.menuOptions = append(m.client.menuOptions, configs...)
-	if len(configs) > 0 {
-		m.client.menuOptions = append(m.client.menuOptions, sessionClientRemove)
-	}
-	m.client.menuOptions = append(m.client.menuOptions, sessionClientAdd)
-	return nil
-}
-
-func (m *configuratorSessionModel) initNameInput() {
-	ti := textinput.New()
-	ti.Prompt = "> "
-	ti.Placeholder = "Give it a name"
-	ti.CharLimit = 256
-	ti.SetWidth(40)
-	ti.SetValue("")
-	ti.Focus()
-	m.client.addNameInput = ti
-}
-
-func (m *configuratorSessionModel) tryFormatJSON() {
-	raw := m.client.addJSONInput.Value()
-	if strings.TrimSpace(raw) == "" {
-		return
-	}
-	var obj json.RawMessage
-	if err := json.Unmarshal([]byte(raw), &obj); err != nil {
-		return
-	}
-	pretty, err := json.MarshalIndent(obj, "", "  ")
-	if err != nil {
-		return
-	}
-	if string(pretty) != raw {
-		m.client.addJSONInput.SetValue(string(pretty))
-	}
-}
-
-func (m *configuratorSessionModel) initJSONInput() {
-	ta := textarea.New()
-	ta.Prompt = "> "
-	ta.Placeholder = "Paste it here"
-	ta.SetWidth(80)
-	ta.SetHeight(10)
-	ta.ShowLineNumbers = true
-	styles := ta.Styles()
-	styles.Focused.CursorLine = styles.Focused.Text
-	ta.SetStyles(styles)
-	ta.SetValue("")
-	ta.Focus()
-	m.client.addJSONInput = ta
-}
-
-func (m *configuratorSessionModel) adjustInputsToViewport() {
-	if m.width <= 0 {
-		return
-	}
-	contentWidth := contentWidthForTerminal(m.width)
-	available := maxInt(1, contentWidth-resolveUIStyles(m.preferences).inputFrame.GetHorizontalFrameSize())
-	m.client.addNameInput.SetWidth(minInt(40, available))
-	m.client.addJSONInput.SetWidth(minInt(80, available))
-	if m.height > 18 {
-		m.client.addJSONInput.SetHeight(m.height - 18)
-	}
-}
-
-func (m configuratorSessionModel) renderSelectionScreen(
-	screenTitle string,
-	notice string,
-	options []string,
-	cursor int,
-	hint string,
-) string {
-	styles := resolveUIStyles(m.preferences)
-	contentWidth := 0
-	if m.width > 0 {
-		contentWidth = contentWidthForTerminal(m.width)
-	}
-
-	rows := renderSelectableRows(options, cursor, contentWidth, styles)
-	body := make([]string, 0, len(rows)+2)
-	if strings.TrimSpace(notice) != "" {
-		body = append(body, notice, "")
-	}
-	body = append(body, rows...)
-	return renderScreen(
-		m.width,
-		m.height,
-		m.tabsLine(styles),
-		screenTitle,
-		body,
-		hint,
-		m.preferences,
-		styles,
-	)
-}
-
-func (m configuratorSessionModel) renderDaemonManageScreen() string {
-	styles := resolveUIStyles(m.preferences)
-	contentWidth := 0
-	if m.width > 0 {
-		contentWidth = contentWidthForTerminal(m.width)
-	}
-
-	rows := renderSelectableRows(m.daemon.menuOptions, m.cursor, contentWidth, styles)
-	body := make([]string, 0, len(rows)+18)
-	body = append(body, styles.title.Render("Daemon Status"))
-	body = append(body, daemonSectionDivider(contentWidth))
-	body = append(body, strings.Split(m.daemonStatusLine(), "\n")...)
-	if !m.daemon.updatedAt.IsZero() {
-		body = append(body, styles.meta.Render("Updated: "+m.daemon.updatedAt.Format("15:04:05")))
-	}
-
-	if notice := strings.TrimSpace(m.notice); notice != "" {
-		body = append(body, "", notice)
-	}
-
-	body = append(body, "", styles.title.Render("Actions"), daemonSectionDivider(contentWidth))
-	body = append(body, rows...)
-
-	return renderScreen(
-		m.width,
-		m.height,
-		m.tabsLine(styles),
-		"Setup/Manage daemon",
-		body,
-		"up/k down/j move | Enter select | Tab switch tabs | Esc back | ctrl+c exit",
-		m.preferences,
-		styles,
-	)
-}
-
-func daemonSectionDivider(contentWidth int) string {
-	if contentWidth <= 0 {
-		return strings.Repeat("-", 24)
-	}
-	return strings.Repeat("-", maxInt(12, minInt(40, contentWidth)))
-}
-
-func daemonMenuCursorAfterRefresh(options []string, selected string, fallbackCursor int) int {
-	if len(options) == 0 {
-		return 0
-	}
-	if selected = strings.TrimSpace(selected); selected != "" {
-		if idx := slices.Index(options, selected); idx >= 0 {
-			return idx
-		}
-	}
-	if fallbackCursor < 0 {
-		return 0
-	}
-	return minInt(fallbackCursor, len(options)-1)
-}
-
-func (m configuratorSessionModel) inputContainerWidth() int {
-	if m.width > 0 {
-		return maxInt(1, contentWidthForTerminal(m.width))
-	}
-	return 40 + resolveUIStyles(m.preferences).inputFrame.GetHorizontalFrameSize()
-}
-
-func (m configuratorSessionModel) settingsTabView() string {
-	styles := resolveUIStyles(m.preferences)
-	contentWidth := 0
-	if m.width > 0 {
-		contentWidth = contentWidthForTerminal(m.width)
-	}
-	body := renderSelectableRows(m.settingsRows(), m.settingsCursor, contentWidth, styles)
-	return renderScreen(
-		m.width,
-		m.height,
-		m.tabsLine(styles),
-		"",
-		body,
-		"up/k down/j row | left/right/Enter change | Tab switch tabs | Esc back | ctrl+c exit",
-		m.preferences,
-		styles,
-	)
-}
-
-func (m configuratorSessionModel) logsTabView() string {
-	styles := resolveUIStyles(m.preferences)
-	body := []string{m.logs.view()}
-	return renderScreen(
-		m.width,
-		m.height,
-		m.tabsLine(styles),
-		"",
-		body,
-		configuratorLogsHint,
-		m.preferences,
-		styles,
-	)
-}
-
-func (m configuratorSessionModel) tabsLine(styles uiStyles) string {
-	contentWidth := contentWidthForTerminal(m.width)
-	return renderTabsLine(productLabel(), "configurator", selectorTabs[:], m.tab, contentWidth, m.preferences.Theme, styles)
-}
-
-func (m configuratorSessionModel) logsFeed() RuntimeLogFeed {
-	return globalRuntimeLogFeed()
-}
-
-func configuratorLogUpdateCmd(feed RuntimeLogFeed, stop <-chan struct{}, seq uint64) tea.Cmd {
-	return logViewportUpdateCmd(feed, stop, seq)
 }
 
 func (m *configuratorSessionModel) updateCursor(keyMsg tea.KeyMsg, listSize int) {
@@ -1822,68 +502,4 @@ func (m *configuratorSessionModel) updateCursor(keyMsg tea.KeyMsg, listSize int)
 			m.cursor++
 		}
 	}
-}
-
-func buildServerManageLabels(peers []appConfiguration.ServerPeer) []string {
-	labels := make([]string, 0, len(peers))
-	for _, peer := range peers {
-		labels = append(labels, serverPeerOptionLabel(peer))
-	}
-	return labels
-}
-
-func serverPeerDisplayName(peer appConfiguration.ServerPeer) string {
-	name := strings.TrimSpace(peer.Name)
-	if name == "" {
-		return fmt.Sprintf("client-%d", peer.ClientID)
-	}
-	return name
-}
-
-func serverPeerOptionLabel(peer appConfiguration.ServerPeer) string {
-	status := "disabled"
-	if peer.Enabled {
-		status = "enabled"
-	}
-	name := serverPeerDisplayName(peer)
-	return fmt.Sprintf("#%d %s [%s]", peer.ClientID, name, status)
-}
-
-func summarizeInvalidConfigurationError(err error) string {
-	if err == nil {
-		return ""
-	}
-
-	message := strings.TrimSpace(err.Error())
-	normalized := strings.ToLower(message)
-	if strings.Contains(normalized, "invalid client configuration (") {
-		if separatorIdx := strings.Index(message, "): "); separatorIdx >= 0 && separatorIdx+3 <= len(message) {
-			message = message[separatorIdx+3:]
-		}
-	}
-	message = strings.Join(strings.Fields(message), " ")
-	if len(message) > 120 {
-		return message[:117] + "..."
-	}
-	return message
-}
-
-func isInvalidClientConfigurationError(err error) bool {
-	if err == nil {
-		return false
-	}
-
-	normalized := strings.ToLower(err.Error())
-	invalidMessages := []string{
-		"invalid client configuration",
-		"invalid character",
-		"cannot unmarshal",
-		"unexpected eof",
-	}
-	for _, message := range invalidMessages {
-		if strings.Contains(normalized, message) {
-			return true
-		}
-	}
-	return false
 }
