+++ src/presentation/ui/tui/internal/bubble_tea/log_viewport_test.go
@@ -22,6 +22,71 @@ func (f *logViewportTestFeed) TailInto(dst []string, limit int) int {
 	return copy(dst, src)
 }
 
+type logViewportTestChangeFeed struct {
+	logViewportTestFeed
+	changes <-chan struct{}
+}
+
+func (f *logViewportTestChangeFeed) Changes() <-chan struct{} {
+	return f.changes
+}
+
+func TestLogViewportUpdateCmd_ReturnsCommand(t *testing.T) {
+	tests := []struct {
+		name string
+		feed RuntimeLogFeed
+	}{
+		{name: "nil feed"},
+		{name: "plain feed", feed: &logViewportTestFeed{}},
+		{name: "change feed", feed: &logViewportTestChangeFeed{changes: make(chan struct{})}},
+		{name: "nil changes", feed: &logViewportTestChangeFeed{}},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if cmd := logViewportUpdateCmd(tt.feed, make(chan struct{}), 5); cmd == nil {
+				t.Fatal("expected non-nil command")
+			}
+		})
+	}
+}
+
+func TestLogViewportUpdateCmd_StopChannelClosed(t *testing.T) {
+	stop := make(chan struct{})
+	close(stop)
+
+	msg := logViewportUpdateCmd(
+		&logViewportTestChangeFeed{changes: make(chan struct{})},
+		stop,
+		7,
+	)()
+	tick, ok := msg.(logViewportTickMsg)
+	if !ok {
+		t.Fatalf("expected logViewportTickMsg, got %T", msg)
+	}
+	if tick.seq != 0 {
+		t.Fatalf("expected seq=0 when stop closed, got %d", tick.seq)
+	}
+}
+
+func TestLogViewportUpdateCmd_ChangesChannelFires(t *testing.T) {
+	changes := make(chan struct{}, 1)
+	changes <- struct{}{}
+
+	msg := logViewportUpdateCmd(
+		&logViewportTestChangeFeed{changes: changes},
+		make(chan struct{}),
+		7,
+	)()
+	tick, ok := msg.(logViewportTickMsg)
+	if !ok {
+		t.Fatalf("expected logViewportTickMsg, got %T", msg)
+	}
+	if tick.seq != 7 {
+		t.Fatalf("expected seq=7 when changes fires, got %d", tick.seq)
+	}
+}
+
 func TestNewLogViewport(t *testing.T) {
 	lv := newLogViewport()
 	if !lv.ready {
