+++ src/presentation/ui/tui/internal/bubble_tea/unified_session_test.go
@@ -29,7 +29,7 @@ func (errReader) Read([]byte) (int, error) {
 }
 
 func defaultUnifiedConfigOpts() ConfiguratorSessionOptions {
-	return sessionOptionsWithControl(defaultSessionConfigurationControl())
+	return testSessionOptions(newTestConfigurationControl())
 }
 
 func newTestUnifiedModel(t *testing.T) (unifiedSessionModel, chan unifiedEvent) {
@@ -611,53 +611,6 @@ func TestUnifiedSession_UpdateRuntime_ReconfigureRequested_DisablesAutoConnectIn
 	updated := result.(unifiedSessionModel)
 	if updated.settings.Preferences().AutoConnect {
 		t.Fatal("expected AutoConnect=false after runtime stop reconfigure in client mode")
-	}
-}
-
-func TestUnifiedSession_UpdateRuntime_ReconfigureRequested_AutoConnectPersistFails_ShowsFatalError(t *testing.T) {
-	prev := persistAutoConnectDisabledToDisk
-	persistAutoConnectDisabledToDisk = func(UIPreferences) error {
-		return errors.New("disk full")
-	}
-	t.Cleanup(func() { persistAutoConnectDisabledToDisk = prev })
-
-	m, events := newTestUnifiedModel(t)
-	m.phase = phaseRuntime
-
-	prefs := m.settings.Preferences()
-	prefs.AutoConnect = true
-	m.settings.update(prefs)
-
-	rt := NewRuntimeDashboard(context.Background(), RuntimeDashboardOptions{}, m.settings)
-	rt.mode = runtime.ModeClient
-	rt.reconfigureRequested = true
-	m.runtime = &rt
-
-	result, cmd := m.Update(tea.KeyPressMsg{Code: 'x', Text: "x"})
-	updated := result.(unifiedSessionModel)
-	if updated.phase != phaseFatalError {
-		t.Fatalf("expected phaseFatalError on AutoConnect persist failure, got %d", updated.phase)
-	}
-	if updated.fatalError == nil {
-		t.Fatal("expected fatalError model to be set")
-	}
-	if !strings.Contains(updated.fatalError.message, "Failed to persist AutoConnect=false before reconfigure") {
-		t.Fatalf("expected fatal error message to mention persistence failure, got %q", updated.fatalError.message)
-	}
-	if !strings.Contains(updated.fatalError.message, "disk full") {
-		t.Fatalf("expected underlying persistence error in fatal message, got %q", updated.fatalError.message)
-	}
-	if cmd != nil {
-		t.Fatal("expected nil cmd when transitioning to fatal error")
-	}
-	if updated.settings.Preferences().AutoConnect {
-		t.Fatal("expected AutoConnect=false in-memory even when persistence fails")
-	}
-
-	select {
-	case event := <-events:
-		t.Fatalf("did not expect runtime event on persistence failure, got %d", event.kind)
-	default:
 	}
 }
